Add tests to verify fieldNames is received correct
A tests has been added to capture the behavior of how entity field names
are taken in code.
